+++ zarrs_storage/CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Bump `unsafe_cell_slice` to 0.2.0
  - **Breaking**: Change `ByteRange::FromEnd` to `ByteRange::Suffix`
  - **Breaking**: implement `AsyncReadableWritableStorageTraits` for `T: AsyncReadableStorageTraits + AsyncWritableStorageTraits`
+ - **Breaking**: Bump MSRV to 1.77 (21 March, 2024)
 
 ### Removed
  - **Breaking**: Remove `ByteRange::offset()`